VW-927 Bug/fix Change datetype model Production process

+++ Service.Manafacturing.Model/ProductionStep/ProductionStepModel.cs
@@ -17,7 +17,7 @@ namespace VErp.Services.Manafacturing.Model.ProductionStep
         public string Title { get; set; }
         public long? ParentId { get; set; }
         public EnumProductionProcess.ContainerType ContainerTypeId { get; set; }
-        public int ContainerId { get; set; }
+        public long ContainerId { get; set; }
         public int SortOrder { get; set; }
         public bool? IsGroup { get; set; }
     }
